+++ src/main/kotlin/dev/m2en/citation/Main.kt
@@ -16,6 +16,7 @@ import dev.kord.gateway.Intents
 import dev.kord.core.event.message.ReactionAddEvent
 import dev.kord.gateway.PrivilegedIntent
 import dev.m2en.citation.command.*
+import dev.m2en.citation.command.chat.HelpCommand
 import io.github.cdimascio.dotenv.dotenv
 
 @OptIn(PrivilegedIntent::class)
