+++ src/main/kotlin/dev/m2en/citation/command/message/DebugCommand.kt
