+++ src/main/kotlin/dev/m2en/citation/command/message/RegisterCommand.kt
